requirements.txt
praw>=7.7.0
plyer>=2.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
